Refactor and clean up of events infrastructure
Core idea: There's only an event node type now in the command
hierarchy, instead of draw, blit, sync, etc types. Can call
GetEventNodeType() to get the exact type. The benefit is that they
are all treated as event nodes now, and there's less of a
need to special-case the different event types.